Finalizado verão MVP para deploy

+++ english_school/app/Http/Controllers/CustomUserController.php
@@ -13,23 +13,49 @@ class CustomUserController extends Controller
 {
     public function index()
     {
+        $search = request('search');
+        $searchTerm = '%'.$search.'%';
+
+        $user = auth()->user();
+
+        if($search)
+        {
+            $users = DB::select('SELECT * FROM CUSTOM_USERS WHERE SCHOOL_ID = ? 
+            AND (NAME LIKE ? OR EMAIL LIKE ? )', [$user->school_id, $searchTerm, $searchTerm]);
+
+            return view('custom_users.dashboard', [
+                'users' => $users,
+                'search' => $search
+            ]);
+        }
+
         $users = DB::select('SELECT * FROM CUSTOM_USERS WHERE SCHOOL_ID = ?', [auth()->user()->school_id]);
         
         return view('custom_users.dashboard', [
-            'users' => $users
+            'users' => $users,
+            'search' => $search
         ]);
     }
 
     public function create()
     {
+        if(auth()->user()->admin == false || auth()->user()->admin == null) {
+            return redirect('/custom_users');
+        }
+
         return view('custom_users.create');
     }
 
     public function store(Request $request)
     {
+        if(auth()->user()->admin == false || auth()->user()->admin == null) {
+            return redirect('/custom_users');
+        }
+
         $user = DB::select(
-            'SELECT * FROM CUSTOM_USERS WHERE email = ?',
-            [$request->email]
+            'SELECT * FROM CUSTOM_USERS WHERE email = ?
+            AND school_id = ?',
+            [$request->email, auth()->user()->school_id]
         );
 
         if(count($user) > 0) {
@@ -41,6 +67,7 @@ class CustomUserController extends Controller
         $user->email = $request->email;
         $user->password = Hash::make($request->password);
         $user->school_id = auth()->user()->school_id;
+        $user->admin = false;
         $user->save();
 
         return redirect('/custom_users')->with('msg', 'Usuário criado com sucesso!');
@@ -48,6 +75,10 @@ class CustomUserController extends Controller
 
     public function edit($id)
     {
+        if(auth()->user()->admin == false || auth()->user()->admin == null) {
+            return redirect('/custom_users');
+        }
+
         $user = CustomUser::findOrFail($id);
 
         return view('custom_users.edit', [
@@ -57,6 +88,10 @@ class CustomUserController extends Controller
 
     public function update(Request $request, $id)
     {
+        if(auth()->user()->admin == false || auth()->user()->admin == null) {
+            return redirect('/custom_users');
+        }
+
         $user = CustomUser::findOrFail($id);
     
         $user->name = $request->name;
@@ -68,7 +103,15 @@ class CustomUserController extends Controller
 
     public function delete($id)
     {
+        if(auth()->user()->admin == false || auth()->user()->admin == null) {
+            return redirect('/custom_users');
+        }
+        
         $user = CustomUser::findOrFail($id);
+
+        if($user->id == auth()->user()->id) {
+            return redirect('/custom_users')->with('msg', 'Você não pode excluir a si mesmo!');
+        }
 
         return view('custom_users.delete', [
             'user' => $user
@@ -77,7 +120,15 @@ class CustomUserController extends Controller
 
     public function destroy($id)
     {
+        if(auth()->user()->admin == false || auth()->user()->admin == null) {
+            return redirect('/custom_users');
+        }
+        
         $user = CustomUser::findOrFail($id);
+
+        if($user->id == auth()->user()->id) {
+            return redirect('/custom_users')->with('msg', 'Você não pode excluir a si mesmo!');
+        }
 
         $user->delete();
 
@@ -87,7 +138,6 @@ class CustomUserController extends Controller
     public function show($id)
     {
         $user = CustomUser::findOrFail($id);
-
 
         return view('custom_users.show', [
             'user' => $user
